Refactor entity handler logic into focused files
